requirements.txt
pandas==1.0.3
numpy==1.18.1
matplotlib==3.1.3
scikit-learn==0.22.1
pickle
tkinter
pillow
scipy
